reorganize form components
