+++ components/CreateCharacterForm.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Button from './Button';
-import Input from './Input';
+import Button from './forms/Button';
+import Input from './forms/Input';
 
 const CreateCharacterForm = (props) => {
 	return (
